Add helper admin scripts (#1)
Add helper scripts and licenses

Instead of having a user install java and download specific kafka
versions to interact with the containers, a bunch of helper scripts were
created to run the admin commands in an interactive container connected
to the kafka server container's network that is exposed to localhost.
This makes testing the various versions easier.  Licensing was also
added.