Fix the first attempt at wrappers
Additional manual tests showed that the wrappers were not working as
expected when operating outside the bin path and redirecting input to
the command.  Made some slight adjustments to remove the psuedo terminal
option from docker, created the kafka-run script and used that to
capture the working diretory and script name, and since everything
looked alike, I'm just linking the scripts instead of duplicating them.